apply multithread to all scene

+++ RayTracing/Head/HitableList.h
@@ -13,7 +13,7 @@ public:
 inline bool HitableList::hit(const Ray& r, float t_min, float t_max, Hit_Record& rec) const {
 	Hit_Record temp_rec;
 	bool hit_anything = false;
-	double closest_so_far = t_max;
+	float closest_so_far = t_max;
 	for (int i = 0; i < list_size; i++) {
 		if (list[i]->hit(r, t_min, closest_so_far, temp_rec)) {
 			hit_anything = true;
