+++ RayTracing/Head/Random.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <cstdlib>
 
-inline double random_double() {
+inline float random_double() {
 	return rand() / (RAND_MAX + 1.0f);
 }
